+++ src/flextype/Support/Parsers/Shortcodes/RawShortcode.php
@@ -12,10 +12,10 @@ use Thunder\Shortcode\Events;
 use Thunder\Shortcode\Shortcode\ShortcodeInterface;
 
 // Shortcode: [raw]
-if (flextype('registry')->get('flextype.settings.shortcode.shortcodes.raw.enabled')) {
-    flextype('shortcode')->addHandler('raw', static function (ShortcodeInterface $s) {
+if (flextype('registry')->get('flextype.settings.parsers.shortcode.shortcodes.raw.enabled')) {
+    flextype('parsers')->shortcode()->addHandler('raw', static function (ShortcodeInterface $s) {
         return $s->getContent();
     });
 
-    flextype('shortcode')->addEventHandler(Events::FILTER_SHORTCODES, new FilterRawEventHandler(['raw']));
+    flextype('parsers')->shortcode()->addEventHandler(Events::FILTER_SHORTCODES, new FilterRawEventHandler(['raw']));
 }