+++ src/flextype/Support/Parsers/Shortcodes/RegistryShortcode.php
@@ -10,8 +10,8 @@ declare(strict_types=1);
 use Thunder\Shortcode\Shortcode\ShortcodeInterface;
 
 // Shortcode: [registry_get name="item-name" default="default-value"]
-if (flextype('registry')->get('flextype.settings.shortcode.shortcodes.registry.enabled')) {
-    flextype('shortcode')->addHandler('registry_get', static function (ShortcodeInterface $s) {
+if (flextype('registry')->get('flextype.settings.parsers.shortcode.shortcodes.registry.enabled')) {
+    flextype('parsers')->shortcode()->addHandler('registry_get', static function (ShortcodeInterface $s) {
         return flextype('registry')->get($s->getParameter('name'), $s->getParameter('default'));
     });
 }